requirements.txt
langchain
langchain-google-genai
langchain-community
faiss-cpu
python-dotenv
pypdf
langchain-text-splitters
langchain-classic
